refactor: proofs_sum helper, 64-bit balances, bounded amount decoding
The proof-amount sum loop was hand-rolled ten times across the console,
display, NFC, and wallet code even though Wallet::balance() existed.
Add cashu::proofs_sum (int64) and route everything through it/balance(),
which is now int64 as well.

Deserializers gain bounds checks: JSON get_int rejects numbers outside
[0, INT32_MAX] (cJSON's valueint saturates silently), and the CBOR
integer reader rejects negatives and 64-bit values it previously
truncated with a bare cast. Widening the model's amount fields to
uint64 proper is deferred — it needs an NVS schema migration (marked
with a TODO on the Proof struct).

+++ main/wallet.hpp
@@ -86,7 +86,7 @@ public:
     bool check_melt_quote(const std::string& quote_id, MeltQuote& quote_out);
     bool melt_tokens(const MeltQuote& quote, int& change_amount);
 
-    int balance() const;
+    int64_t balance() const;
     bool select_proofs(int amount_needed, std::vector<Proof>& selected,
                        std::vector<Proof>& remaining);
